(refactor): extracted union shape handling out of MinShapeAlgorithm to work in isolation.
